requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.3.6
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.0
defusedxml==0.7.1
Django==4.2
django-allauth==0.57.0
django-browser-reload==1.12.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
oauthlib==3.2.2
packaging==24.2
Pillow==10.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
redis==5.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
